fix: Resolve CI failures in PR #68
- Fix clang-format alignment of inline comments in MemoryPoolStats
- Fix implicit widening conversion in kDefaultPoolSize constant (use 256ULL)
- Fix type conversion warning in test_memory_pool.cpp (explicit cast to double)
- Fix include-what-you-use issues:
  - Use forward declaration for DeviceAllocator in memory_pool.hpp
  - Move allocator.hpp include to memory_pool.cpp
  - Add missing includes: <new>, <utility>

+++ tests/test_memory_pool.cpp
@@ -190,7 +190,7 @@ TEST_F(MemoryPoolTest, PerformanceBenchmark) {
   if (duration_pool.count() > 0) {
     std::cout << "Speedup: "
               << static_cast<double>(duration_direct.count()) /
-                     duration_pool.count()
+                     static_cast<double>(duration_pool.count())
               << "x\n";
   }
